requirements.txt
streamlit
PyPDF2
sentence-transformers
pinecone-client
cohere
